requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
docx==0.2.4
filelock==3.16.0
filetype==1.2.0
Flask==3.1.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-deepseek==0.1.3
langchain-google-genai==2.1.1
langchain-ollama==0.3.0
langchain-openai==0.3.10
langchain-text-splitters==0.3.8
langsmith==0.3.18
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.3.2
mypy-extensions==1.0.0
numpy==2.2.4
ollama==0.4.7
openai==1.68.2
orjson==3.10.16
packaging==24.2
pillow==11.1.0
platformdirs==4.3.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.39
tavily-python==0.5.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.26.4
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
